* Fix more 64 bit compat errors in KAI, also dont change signedness of casts in micropather,   not sure whether the specific (un)signedness is needed for proper functionality.
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@2413 37977431-3df6-0310-b722-df95706aa16b

+++ AI/Global/KAI 0.11/PathFinder.cpp
@@ -307,7 +307,7 @@ void* CPathFinder::XY2Node(int x, int y)
 
 void CPathFinder::Node2XY(void* node, int* x, int* y)
 {
-	int index = (int)node;
+	long index = (long)node;
 	*y = index / PathMapXSize;
 	*x = index - *y * PathMapXSize;
 }
@@ -316,7 +316,7 @@ float3 CPathFinder::Node2Pos(void*node)
 {
 	float3 pos;
 	int multiplier = int(8 * resmodifier);
-	int index = (int)node;
+	long index = (long)node;
 	pos.z = (index / PathMapXSize) * multiplier; /////////////////////OPTiMIZE
 	pos.x = (index - ((index / PathMapXSize) * PathMapXSize)) * multiplier;
 	return pos;
